+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "Cursor Insights" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-# Unreleased
+# 1.1.0
 
 ### Changed